fixing RANDCOLOR.py colors

+++ program/RANDCOLOR.py
@@ -21,25 +21,25 @@ def rand_color(strip, color):
     random_place = random.randrange(strip.numPixels())
     for j in range(2,-1,-1):
         strip.setPixelColor(random_place - j,
-                            Color(color[0] / (j + 1),
-                                  color[1] / (j + 1),
-                                  color[2] / (j + 1)))
+                            Color(int(color[0] / (j + 1)),
+                                  int(color[1] / (j + 1)),
+                                  int(color[2] / (j + 1))))
         strip.setPixelColor(random_place+j,
-                            Color(color[0] / (j + 1),
-                                  color[1] / (j + 1),
-                                  color[2] / (j + 1)))
+                            Color(int(color[0] / (j + 1)),
+                                  int(color[1] / (j + 1)),
+                                  int(color[2] / (j + 1))))
         strip.setBrightness(int(255/j+1))
         strip.show()
         time.sleep(100 / 1000.0)
     for j in range(2, -1, -1):
         strip.setPixelColor(random_place - j,
-                            Color(color[0] - color[0] / (j + 1),
-                                  color[1] - color[1] / (j + 1),
-                                  color[2] - color[2] / (j + 1)))
+                            Color(int(color[0] - color[0] / (j + 1)),
+                                  int(color[1] - color[1] / (j + 1)),
+                                  int(color[2] - color[2] / (j + 1))))
         strip.setPixelColor(random_place + j,
-                            Color(color[0] - color[0] / (j + 1),
-                                  color[1] - color[1] / (j + 1),
-                                  color[2] - color[2] / (j + 1)))
+                            Color(int(color[0] - color[0] / (j + 1)),
+                                  int(color[1] - color[1] / (j + 1)),
+                                  int(color[2] - color[2] / (j + 1))))
         strip.setBrightness(int(255-255/(j+1)))
         strip.show()
         time.sleep(100 / 1000.0)
